requirements.txt
openai
astunparse
GitPython
pyyaml